[ML-187] Support spark 3.1.3 and 3.2.0 and support CDH (#197)
* support spark 3.1.3


* support spark 3.2.1


* support CDH spark


* update
