+++ mllib-dal/src/main/scala/com/intel/oap/mllib/Utils.scala
@@ -19,8 +19,7 @@ package com.intel.oap.mllib
 import org.apache.spark.ml.linalg.Vector
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.SparkSession
-import org.apache.spark.{SparkConf, SparkContext}
-
+import org.apache.spark.{SPARK_VERSION, SparkConf, SparkContext}
 import java.net.InetAddress
 
 object Utils {
@@ -154,5 +153,17 @@ object Utils {
 
     // Return executor number (exclude driver)
     executorInfos.length - 1
+  }
+  def getSparkVersion(): String = {
+    // For example: CHD spark version is 3.1.1.3.1.7290.5-2.
+    // The string before the third dot is the spark version.
+    val array = SPARK_VERSION.split("\\.")
+    val sparkVersion = if (array.size > 3) {
+      val version = array.take(3).mkString(".")
+      version
+    } else {
+      SPARK_VERSION
+    }
+    sparkVersion
   }
 }
