+++ mllib-dal/src/main/scala/com/intel/oap/mllib/classification/NaiveBayesShim.scala
@@ -16,9 +16,11 @@
 
 package com.intel.oap.mllib.classification
 
+import com.intel.oap.mllib.Utils
+
 import org.apache.spark.internal.Logging
 import org.apache.spark.ml.classification.NaiveBayesModel
-import org.apache.spark.ml.classification.spark320.{NaiveBayes => NaiveBayesSpark320}
+import org.apache.spark.ml.classification.spark321.{NaiveBayes => NaiveBayesSpark321}
 import org.apache.spark.ml.param.ParamMap
 import org.apache.spark.sql.Dataset
 import org.apache.spark.{SPARK_VERSION, SparkException}
@@ -31,8 +33,9 @@ trait NaiveBayesShim extends Logging {
 object NaiveBayesShim extends Logging {
   def create(uid: String): NaiveBayesShim = {
     logInfo(s"Loading NaiveBayes for Spark $SPARK_VERSION")
-    val shim = SPARK_VERSION match {
-      case "3.1.1" | "3.1.2" | "3.2.0" => new NaiveBayesSpark320(uid)
+
+    val shim = Utils.getSparkVersion() match {
+      case "3.1.1" | "3.1.2" | "3.1.3" | "3.2.0" | "3.2.1" => new NaiveBayesSpark321(uid)
       case _ => throw new SparkException(s"Unsupported Spark version $SPARK_VERSION")
     }
     shim
