+++ mllib-dal/src/main/scala/com/intel/oap/mllib/clustering/KMeansShim.scala
@@ -16,9 +16,11 @@
 
 package com.intel.oap.mllib.clustering
 
+import com.intel.oap.mllib.Utils
+
 import org.apache.spark.internal.Logging
 import org.apache.spark.ml.clustering.{KMeans, KMeansModel}
-import org.apache.spark.ml.clustering.spark320.{KMeans => KMeansSpark320}
+import org.apache.spark.ml.clustering.spark321.{KMeans => KMeansSpark321}
 import org.apache.spark.ml.param.ParamMap
 import org.apache.spark.sql.Dataset
 import org.apache.spark.{SPARK_VERSION, SparkException}
@@ -31,8 +33,8 @@ trait KMeansShim extends Logging {
 object KMeansShim extends Logging {
   def create(uid: String): KMeansShim = {
     logInfo(s"Loading KMeans for Spark $SPARK_VERSION")
-    val kmeans = SPARK_VERSION match {
-      case "3.1.1" | "3.1.2" | "3.2.0" => new KMeansSpark320(uid)
+    val kmeans = Utils.getSparkVersion() match {
+      case "3.1.1" | "3.1.2" | "3.1.3" | "3.2.0" | "3.2.1" => new KMeansSpark321(uid)
       case _ => throw new SparkException(s"Unsupported Spark version $SPARK_VERSION")
     }
     kmeans
