+++ mllib-dal/src/main/scala/com/intel/oap/mllib/regression/LinearRegressionShim.scala
@@ -16,11 +16,13 @@
 
 package com.intel.oap.mllib.regression
 
+import com.intel.oap.mllib.Utils
+
 import org.apache.spark.internal.Logging
 import org.apache.spark.ml.param.ParamMap
 import org.apache.spark.ml.regression.LinearRegressionModel
-import org.apache.spark.ml.regression.spark312.{LinearRegression => LinearRegressionSpark312}
-import org.apache.spark.ml.regression.spark320.{LinearRegression => LinearRegressionSpark320}
+import org.apache.spark.ml.regression.spark313.{LinearRegression => LinearRegressionSpark313}
+import org.apache.spark.ml.regression.spark321.{LinearRegression => LinearRegressionSpark321}
 import org.apache.spark.sql.Dataset
 import org.apache.spark.{SPARK_VERSION, SparkException}
 
@@ -32,9 +34,9 @@ trait LinearRegressionShim extends Serializable with Logging {
 object LinearRegressionShim extends Logging {
   def create(uid: String): LinearRegressionShim = {
     logInfo(s"Loading ALS for Spark $SPARK_VERSION")
-    val linearRegression = SPARK_VERSION match {
-      case "3.1.1" | "3.1.2" => new LinearRegressionSpark312(uid)
-      case "3.2.0" => new LinearRegressionSpark320(uid)
+    val linearRegression = Utils.getSparkVersion() match {
+      case "3.1.1" | "3.1.2" | "3.1.3" => new LinearRegressionSpark313(uid)
+      case "3.2.0" | "3.2.1" => new LinearRegressionSpark321(uid)
       case _ => throw new SparkException(s"Unsupported Spark version $SPARK_VERSION")
     }
     linearRegression
