+++ mllib-dal/src/main/scala/org/apache/spark/ml/classification/spark321/NaiveBayes.scala
@@ -17,7 +17,7 @@
  */
 // scalastyle:on
 
-package org.apache.spark.ml.classification.spark320
+package org.apache.spark.ml.classification.spark321
 
 import com.intel.oap.mllib.Utils
 import com.intel.oap.mllib.classification.{NaiveBayesDALImpl, NaiveBayesShim}
