+++ mllib-dal/src/main/scala/org/apache/spark/ml/clustering/spark321/KMeans.scala
@@ -17,7 +17,7 @@
  */
 // scalastyle:on
 
-package org.apache.spark.ml.clustering.spark320
+package org.apache.spark.ml.clustering.spark321
 
 import com.intel.oap.mllib.Utils
 import com.intel.oap.mllib.clustering.{KMeansDALImpl, KMeansShim}
