+++ mllib-dal/src/main/scala/org/apache/spark/ml/feature/spark321/PCA.scala
@@ -17,7 +17,7 @@
  */
 // scalastyle:on
 
-package org.apache.spark.ml.feature.spark320
+package org.apache.spark.ml.feature.spark321
 
 import com.intel.oap.mllib.Utils
 import com.intel.oap.mllib.feature.{PCADALImpl, PCAShim}
