+++ mllib-dal/src/main/scala/org/apache/spark/ml/recommendation/spark321/ALS.scala
@@ -17,7 +17,7 @@
  */
 // scalastyle:on
 
-package org.apache.spark.ml.recommendation.spark320
+package org.apache.spark.ml.recommendation.spark321
 
 import com.github.fommil.netlib.BLAS.{getInstance => blas}
 import com.intel.oap.mllib.{Utils => DALUtils}
