+++ mllib-dal/src/main/scala/org/apache/spark/ml/regression/spark313/LinearRegression.scala
@@ -17,7 +17,7 @@
  */
 // scalastyle:on
 
-package org.apache.spark.ml.regression.spark320
+package org.apache.spark.ml.regression.spark313
 
 import breeze.linalg.{DenseVector => BDV}
 import breeze.optimize.{
@@ -39,7 +39,8 @@ import org.apache.spark.annotation.Since
 import org.apache.spark.internal.Logging
 import org.apache.spark.ml.{PipelineStage, PredictorParams}
 import org.apache.spark.ml.feature._
-import org.apache.spark.ml.linalg.{BLAS, Vector, Vectors}
+import org.apache.spark.ml.linalg.{Vector, Vectors}
+import org.apache.spark.ml.linalg.BLAS._
 import org.apache.spark.ml.optim.WeightedLeastSquares
 import org.apache.spark.ml.optim.aggregator._
 import org.apache.spark.ml.optim.loss.{L2Regularization, RDDLossFunction}
@@ -192,7 +193,7 @@ private[regression] trait LinearRegressionParams extends PredictorParams
  * Note: Fitting with huber loss only supports none and L2 regularization.
  */
 @Since("1.3.0")
-class LinearRegression @Since("1.3.0") (@Since("1.3.0") override val uid: String)
+class LinearRegression @Since("1.3") (@Since("1.3.0") override val uid: String)
   extends SparkLinearRegression with LinearRegressionShim {
 
   import LinearRegression._
@@ -393,8 +394,8 @@ class LinearRegression @Since("1.3.0") (@Since("1.3.0") override val uid: String
     val featuresMean = summarizer.mean.toArray
     val featuresStd = summarizer.std.toArray
 
-    if (!$(fitIntercept) &&
-      (0 until numFeatures).exists(i => featuresStd(i) == 0.0 && featuresMean(i) != 0.0)) {
+    if (!$(fitIntercept) && (0 until numFeatures).exists { i =>
+      featuresStd(i) == 0.0 && featuresMean(i) != 0.0 }) {
       instr.logWarning("Fitting LinearRegressionModel without intercept on dataset with " +
         "constant nonzero column, Spark MLlib outputs zero coefficients for constant nonzero " +
         "columns. This behavior is the same as R glmnet but different from LIBSVM.")
@@ -419,17 +420,17 @@ class LinearRegression @Since("1.3.0") (@Since("1.3.0") override val uid: String
     val optimizer = createOptimizer(effectiveRegParam, effectiveL1RegParam,
       numFeatures, featuresStd)
 
-    val initialSolution = $(loss) match {
+    val initialValues = $(loss) match {
       case SquaredError =>
-        Array.ofDim[Double](numFeatures)
+        Vectors.zeros(numFeatures)
       case Huber =>
         val dim = if ($(fitIntercept)) numFeatures + 2 else numFeatures + 1
-        Array.fill(dim)(1.0)
+        Vectors.dense(Array.fill(dim)(1.0))
     }
 
     val (parameters, objectiveHistory) =
       trainImpl(instances, actualBlockSizeInMB, yMean, yStd,
-        featuresMean, featuresStd, initialSolution, regularization, optimizer)
+        featuresMean, featuresStd, initialValues, regularization, optimizer)
 
     if (parameters == null) {
       val msg = s"${optimizer.getClass.getName} failed."
@@ -581,48 +582,35 @@ class LinearRegression @Since("1.3.0") (@Since("1.3.0") override val uid: String
       yStd: Double,
       featuresMean: Array[Double],
       featuresStd: Array[Double],
-      initialSolution: Array[Double],
+      initialValues: Vector,
       regularization: Option[L2Regularization],
       optimizer: FirstOrderMinimizer[BDV[Double], DiffFunction[BDV[Double]]]) = {
-    val numFeatures = featuresStd.length
-    val inverseStd = featuresStd.map(std => if (std != 0) 1.0 / std else 0.0)
-    val scaledMean = Array.tabulate(numFeatures)(i => inverseStd(i) * featuresMean(i))
-    val bcInverseStd = instances.context.broadcast(inverseStd)
-    val bcScaledMean = instances.context.broadcast(scaledMean)
+    val bcFeaturesMean = instances.context.broadcast(featuresMean)
+    val bcFeaturesStd = instances.context.broadcast(featuresStd)
 
     val standardized = instances.mapPartitions { iter =>
-      val func = StandardScalerModel.getTransformFunc(Array.empty, bcInverseStd.value, false, true)
+      val inverseStd = bcFeaturesStd.value.map { std => if (std != 0) 1.0 / std else 0.0 }
+      val func = StandardScalerModel.getTransformFunc(Array.empty, inverseStd, false, true)
       iter.map { case Instance(label, weight, vec) => Instance(label, weight, func(vec)) }
     }
 
     val maxMemUsage = (actualBlockSizeInMB * 1024L * 1024L).ceil.toLong
     val blocks = InstanceBlock.blokifyWithMaxMemUsage(standardized, maxMemUsage)
       .persist(StorageLevel.MEMORY_AND_DISK)
-      .setName(s"$uid: training blocks (blockSizeInMB=$actualBlockSizeInMB)")
-
-    if ($(fitIntercept) && $(loss) == Huber) {
-      // orginal `initialSolution` is for problem:
-      // y = f(w1 * x1 / std_x1, w2 * x2 / std_x2, ..., intercept)
-      // we should adjust it to the initial solution for problem:
-      // y = f(w1 * (x1 - avg_x1) / std_x1, w2 * (x2 - avg_x2) / std_x2, ..., intercept)
-      // NOTE: this is NOOP before we finally support model initialization
-      val adapt = BLAS.javaBLAS.ddot(numFeatures, initialSolution, 1, scaledMean, 1)
-      initialSolution(numFeatures) += adapt
-    }
+      .setName(s"training blocks (blockSizeInMB=$actualBlockSizeInMB)")
 
     val costFun = $(loss) match {
       case SquaredError =>
-        val getAggregatorFunc = new LeastSquaresBlockAggregator(bcInverseStd, bcScaledMean,
-          $(fitIntercept), yStd, yMean)(_)
+        val getAggregatorFunc = new BlockLeastSquaresAggregator(yStd, yMean, $(fitIntercept),
+          bcFeaturesStd, bcFeaturesMean)(_)
         new RDDLossFunction(blocks, getAggregatorFunc, regularization, $(aggregationDepth))
       case Huber =>
-        val getAggregatorFunc = new HuberBlockAggregator(bcInverseStd, bcScaledMean,
-          $(fitIntercept), $(epsilon))(_)
+        val getAggregatorFunc = new BlockHuberAggregator($(fitIntercept), $(epsilon))(_)
         new RDDLossFunction(blocks, getAggregatorFunc, regularization, $(aggregationDepth))
     }
 
     val states = optimizer.iterations(new CachedDiffFunction(costFun),
-      new BDV(initialSolution))
+      initialValues.asBreeze.toDenseVector)
 
     /*
        Note that in Linear Regression, the objective history (loss + regularization) returned
@@ -642,23 +630,14 @@ class LinearRegression @Since("1.3.0") (@Since("1.3.0") override val uid: String
     }
 
     blocks.unpersist()
-    bcInverseStd.destroy()
-    bcScaledMean.destroy()
+    bcFeaturesMean.destroy()
+    bcFeaturesStd.destroy()
 
-    val solution = if (state == null) null else state.x.toArray
-    if ($(fitIntercept) && $(loss) == Huber && solution != null) {
-      // the final solution is for problem:
-      // y = f(w1 * (x1 - avg_x1) / std_x1, w2 * (x2 - avg_x2) / std_x2, ..., intercept)
-      // we should adjust it back for original problem:
-      // y = f(w1 * x1 / std_x1, w2 * x2 / std_x2, ..., intercept)
-      val adapt = BLAS.javaBLAS.ddot(numFeatures, solution, 1, scaledMean, 1)
-      solution(numFeatures) -= adapt
-    }
-    (solution, arrayBuilder.result)
+    (if (state == null) null else state.x.toArray, arrayBuilder.result)
   }
 
   private def createModel(
-      solution: Array[Double],
+      parameters: Array[Double],
       yMean: Double,
       yStd: Double,
       featuresMean: Array[Double],
@@ -668,9 +647,20 @@ class LinearRegression @Since("1.3.0") (@Since("1.3.0") override val uid: String
        The coefficients are trained in the scaled space; we're converting them back to
        the original space.
      */
-    val multiplier = if ($(loss) == Huber) 1.0 else yStd
-    val rawCoefficients = Array.tabulate(numFeatures) { i =>
-      if (featuresStd(i) != 0) solution(i) * multiplier / featuresStd(i) else 0.0
+    val rawCoefficients = $(loss) match {
+      case SquaredError => parameters.clone()
+      case Huber => parameters.take(numFeatures)
+    }
+
+    var i = 0
+    val len = rawCoefficients.length
+    val multiplier = $(loss) match {
+      case SquaredError => yStd
+      case Huber => 1.0
+    }
+    while (i < len) {
+      rawCoefficients(i) *= { if (featuresStd(i) != 0.0) multiplier / featuresStd(i) else 0.0 }
+      i += 1
     }
 
     val intercept = if ($(fitIntercept)) {
@@ -681,13 +671,18 @@ class LinearRegression @Since("1.3.0") (@Since("1.3.0") override val uid: String
           after the coefficients are converged. See the following discussion for detail.
           http://stats.stackexchange.com/questions/13617/how-is-the-intercept-computed-in-glmnet
           */
-          yMean - BLAS.dot(Vectors.dense(rawCoefficients), Vectors.dense(featuresMean))
-        case Huber => solution(numFeatures)
+          yMean - dot(Vectors.dense(rawCoefficients), Vectors.dense(featuresMean))
+        case Huber => parameters(numFeatures)
       }
     } else 0.0
 
+    val scale = $(loss) match {
+      case SquaredError => 1.0
+      case Huber => parameters.last
+    }
+
     val coefficients = Vectors.dense(rawCoefficients).compressed
-    val scale = if ($(loss) == Huber) solution.last else 1.0
+
     copyValues(new LinearRegressionModel(uid, coefficients, intercept, scale))
   }
 
@@ -731,6 +726,7 @@ object LinearRegression extends DefaultParamsReadable[LinearRegression] {
   private[regression] val supportedLosses = Array(SquaredError, Huber)
 }
 
+
 /** A writer for LinearRegression that handles the "internal" (or default) format */
 private class InternalLinearRegressionModelWriter
   extends MLWriterFormat with MLFormatRegister {
@@ -771,45 +767,5 @@ private class PMMLLinearRegressionModelWriter
     val oldModel = new OldLinearRegressionModel(instance.coefficients, instance.intercept)
     // Save PMML
     oldModel.toPMML(sc, path)
-  }
-}
-
-@Since("1.6.0")
-object LinearRegressionModel extends MLReadable[LinearRegressionModel] {
-
-  @Since("1.6.0")
-  override def read: MLReader[LinearRegressionModel] = new LinearRegressionModelReader
-
-  @Since("1.6.0")
-  override def load(path: String): LinearRegressionModel = super.load(path)
-
-  private class LinearRegressionModelReader extends MLReader[LinearRegressionModel] {
-
-    /** Checked against metadata when loading model */
-    private val className = classOf[LinearRegressionModel].getName
-
-    override def load(path: String): LinearRegressionModel = {
-      val metadata = DefaultParamsReader.loadMetadata(path, sc, className)
-
-      val dataPath = new Path(path, "data").toString
-      val data = sparkSession.read.format("parquet").load(dataPath)
-      val (majorVersion, minorVersion) = majorMinorVersion(metadata.sparkVersion)
-      val model = if (majorVersion < 2 || (majorVersion == 2 && minorVersion <= 2)) {
-        // Spark 2.2 and before
-        val Row(intercept: Double, coefficients: Vector) =
-          MLUtils.convertVectorColumnsToML(data, "coefficients")
-            .select("intercept", "coefficients")
-            .head()
-        new LinearRegressionModel(metadata.uid, coefficients, intercept)
-      } else {
-        // Spark 2.3 and later
-        val Row(intercept: Double, coefficients: Vector, scale: Double) =
-          data.select("intercept", "coefficients", "scale").head()
-        new LinearRegressionModel(metadata.uid, coefficients, intercept, scale)
-      }
-
-      metadata.getAndSetParams(model)
-      model
-    }
   }
 }
